+++ modules/testing/js/src/main/scala/JsInstances.scala
@@ -1,5 +1,6 @@
 package com.azavea.stac4s.testing
 
+import com.azavea.stac4s.extensions.layer.StacLayer
 import com.azavea.stac4s.geometry.Geometry.{MultiPolygon, Point2d, Polygon}
 import com.azavea.stac4s.geometry._
 import com.azavea.stac4s.types.TemporalExtent
@@ -118,6 +119,17 @@ trait JsInstances {
       Gen.const(().asJsonObject)
     ).mapN(StacCollection.apply)
 
+  private[testing] def stacLayerGen: Gen[StacLayer] = (
+    nonEmptyAlphaRefinedStringGen,
+    TestInstances.bboxGen,
+    geometryGen,
+    TestInstances.stacLayerPropertiesGen,
+    Gen.listOfN(8, TestInstances.stacLinkGen),
+    Gen.const("Feature")
+  ).mapN(
+    StacLayer.apply
+  )
+
   implicit val arbItem: Arbitrary[StacItem] = Arbitrary { stacItemGen }
 
   val arbItemShort: Arbitrary[StacItem] = Arbitrary { stacItemShortGen }
@@ -133,6 +145,10 @@ trait JsInstances {
   implicit val arbGeometry: Arbitrary[Geometry] = Arbitrary { geometryGen }
 
   val arbCollectionShort: Arbitrary[StacCollection] = Arbitrary { stacCollectionShortGen }
+
+  implicit val arbStacLayer: Arbitrary[StacLayer] = Arbitrary {
+    stacLayerGen
+  }
 }
 
 object JsInstances extends JsInstances {}